# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenG2P Approval Workflow Engine",
    "description": "Generic, configurable multi-stage approval workflow engine for OpenG2P modules. Caller services post artifacts; AWE resolves stages and approvers, then notifies the caller via signed webhook callbacks when state changes.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/awe/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List tasks \u2014 by assignee (default = me) and/or by request_id",
        "operationId": "list_tasks_v1_awe_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by assignee. Default `me` resolves to the token's `sub`. Pass `*` (or any non-`me` value) plus `request_id` to enumerate all tasks for a given request \u2014 used by the admin Request Detail page.",
              "default": "me",
              "title": "Assignee"
            },
            "description": "Filter by assignee. Default `me` resolves to the token's `sub`. Pass `*` (or any non-`me` value) plus `request_id` to enumerate all tasks for a given request \u2014 used by the admin Request Detail page."
          },
          {
            "name": "request_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Request Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskOut"
                  },
                  "title": "Response List Tasks V1 Awe Tasks Get"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/awe/tasks/{task_id}/claim": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Claim a task (intent-to-act marker; not required for decision)",
        "operationId": "claim_task_v1_awe_tasks__task_id__claim_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Task is not assigned to the caller and the caller is not an admin.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Task is not assigned to you"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No task with the given id.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-004",
                      "message": "Task not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource is in a state that disallows this transition. Examples: request is already in a terminal state (`approved`/`rejected`/`cancelled`); task is not `open`/`claimed`; policy version is not `draft` for editing; delivery has already succeeded.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-007",
                      "message": "Request is in terminal state 'approved' \u2014 cannot cancel"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/awe/tasks/{task_id}/decision": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Record a decision (approve / reject / abstain) on a task",
        "operationId": "decide_v1_awe_tasks__task_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Task is not assigned to the caller and the caller is not an admin.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Task is not assigned to you"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No approval request with the given id.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-003",
                      "message": "Request not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource is in a state that disallows this transition. Examples: request is already in a terminal state (`approved`/`rejected`/`cancelled`); task is not `open`/`claimed`; policy version is not `draft` for editing; delivery has already succeeded.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-007",
                      "message": "Request is in terminal state 'approved' \u2014 cannot cancel"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/awe/tasks/{task_id}/reassign": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Reassign an open task to a different user (admin only)",
        "operationId": "reassign_v1_awe_tasks__task_id__reassign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignTaskIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            }
          },
          "401": {
            "description": "Bearer token missing, malformed, or fails signature/expiry checks.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Invalid or missing bearer token"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Token is valid but lacks the `AWE_ADMIN` role required for this mutation.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-008",
                      "message": "Caller lacks AWE_ADMIN role"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No approval request with the given id.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-003",
                      "message": "Request not found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource is in a state that disallows this transition. Examples: request is already in a terminal state (`approved`/`rejected`/`cancelled`); task is not `open`/`claimed`; policy version is not `draft` for editing; delivery has already succeeded.",
            "content": {
              "application/json": {
                "example": {
                  "id": "openg2p.awe",
                  "version": "1.0",
                  "responsetime": "2026-04-23T10:00:00.000Z",
                  "errors": [
                    {
                      "errorCode": "AWE-007",
                      "message": "Request is in terminal state 'approved' \u2014 cannot cancel"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "DecisionIn": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action",
            "examples": [
              "approve"
            ]
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "attachments_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments Ref"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "DecisionIn"
      },
      "DecisionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "stage_order": {
            "type": "integer",
            "title": "Stage Order"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "attachments_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments Ref"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "request_id",
          "task_id",
          "stage_order",
          "actor",
          "action",
          "created_at"
        ],
        "title": "DecisionOut"
      },
...
      "ErrorDetail": {
        "properties": {
          "errorCode": {
            "type": "string",
            "title": "Errorcode",
            "description": "OpenG2P-assigned error code. AWE catalog: `AWE-001` (policy not found), `AWE-002` (policy conflict / version clash), `AWE-003` (request not found), `AWE-004` (task not found), `AWE-005` (service not ready \u2014 startup incomplete), `AWE-006` (database health check failed), `AWE-007` (invalid state transition), `AWE-008` (unauthorized / forbidden), `AWE-009` (idempotency key conflict with different payload), `AWE-010` (validation \u2014 bad policy definition).",
            "examples": [
              "AWE-003"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Approval request not found"
            ]
          }
        },
        "type": "object",
        "required": [
          "errorCode",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": "openg2p.awe",
            "examples": [
              "openg2p.awe"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0",
            "examples": [
              "1.0"
            ]
          },
          "responsetime": {
            "type": "string",
            "title": "Responsetime",
            "examples": [
              "2026-04-23T10:00:00.000Z"
            ]
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "responsetime",
          "errors"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReassignTaskIn": {
        "properties": {
          "new_assignee": {
            "type": "string",
            "minLength": 1,
            "title": "New Assignee",
            "examples": [
              "u-bob"
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "new_assignee"
        ],
        "title": "ReassignTaskIn"
      },
...
      "TaskOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "stage_id": {
            "type": "string",
            "title": "Stage Id"
          },
          "stage_order": {
            "type": "integer",
            "title": "Stage Order"
          },
          "assignee": {
            "type": "string",
            "title": "Assignee"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "approver"
          },
          "delegated_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delegated From"
          },
          "reassigned_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reassigned From"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "claimed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "decision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "request_id",
          "stage_id",
          "stage_order",
          "assignee",
          "status",
          "created_at"
        ],
        "title": "TaskOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}